requirements.txt
alembic
aniso8601
attrs
bcrypt
blinker
cachelib
certifi
charset-normalizer
click
colorama
defusedxml
Deprecated
dnspython
email-validator
fillpdf
Flask
Flask-Admin
Flask-Bcrypt
Flask-CKEditor
Flask-Cors
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Mail
Flask-Migrate
Flask-RESTful
flask-restx
Flask-Session
Flask-SQLAlchemy
Flask-WTF
greenlet
gunicorn
httpie
idna
imagekitio
importlib-resources
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
limits
Mako
markdown-it-py
MarkupSafe
mdurl
multidict
ordered-set
packaging
pdf2image
pdfrw2
Pillow
psycopg2
psycopg2-binary
Pygments
PyJWT
PyMuPDF
PyMuPDFb
PySocks
python-dotenv
pytz
redis
referencing
requests
requests-toolbelt
rich
rpds-py
rq
six
SQLAlchemy
typing_extensions
urllib3
Werkzeug
wrapt
WTForms
